backend/requirements.txt
Flask==2.1.2
psycopg2==2.9.3
Flask-Cors==5.0.1
pandas
numpy
xgboost
scikit-learn
sqlalchemy
python-dotenv
seaborn
matplotlib
fastapi
pydantic


